requirements.txt
blinker==1.8.2
click==8.1.8
dnspython==2.6.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
fsspec==2024.12.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
pillow==10.4.0
pymongo==4.10.1
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
Werkzeug==3.0.6
zipp==3.20.2
gunicorn
python-dotenv
